+++ bip_utils/utils/__init__.py
@@ -1,11 +1,7 @@
-from bip_utils.utils.algo import AlgoUtils
-from bip_utils.utils.base32 import Base32Encoder
-from bip_utils.utils.bit import BitUtils
-from bip_utils.utils.coin_names import CoinNames
-from bip_utils.utils.conversion import ConvUtils
-from bip_utils.utils.crypto import CryptoUtils
-from bip_utils.utils.data_bytes import DataBytes
-from bip_utils.utils.entropy_generator import EntropyGenerator
+from bip_utils.utils.conf import CoinNames
+from bip_utils.utils.misc import (
+    AlgoUtils, Base32Encoder, BitUtils, ConvUtils, CryptoUtils, DataBytes
+)
 from bip_utils.utils.mnemonic import (
-    Mnemonic, MnemonicWordsList, MnemonicWordsListFileReader, MnemonicWordsListGetterBase
+    EntropyGenerator, Mnemonic, MnemonicWordsList, MnemonicWordsListFileReader, MnemonicWordsListGetterBase
 )
