+++ bip_utils/utils/conf/__init__.py
@@ -1,0 +1,1 @@
+from bip_utils.utils.conf.coin_names import CoinNames
