+++ bip_utils/utils/conf/coin_names.py
