+++ bip_utils/utils/misc/algo.py
