+++ bip_utils/utils/misc/bit.py
