+++ bip_utils/utils/mnemonic/__init__.py
@@ -1,0 +1,4 @@
+from bip_utils.utils.mnemonic.entropy_generator import EntropyGenerator
+from bip_utils.utils.mnemonic.mnemonic import (
+    Mnemonic, MnemonicWordsList, MnemonicWordsListFileReader, MnemonicWordsListGetterBase
+)
